server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
pymysql>=1.1.0   # MySQL驱动
cryptography>=41.0.0  # 用于PyMySQL的加密支持
python-multipart>=0.0.6
alembic>=1.12.1  # 数据库迁移工具 